requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==1.0.3
google-cloud-aiplatform==1.36.4
google-generativeai==0.3.2
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
python-jose==3.3.0
requests==2.31.0
geopy==2.4.0 